+++ deltafi-ui/CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-
 ### Changed
 
 - Added Pretty Print to content viewer for JSON and XML (#84)
@@ -16,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Message on flow pages if no flow is found (#176)
 - Changed Divider color on flows page (#177)
 - Moved Flows config to a dialog box on Flows page (#168)
+- Added auto select on right click for errors table (#161)
 
 ## [0.95.2]